+++ CHANGES.rst
@@ -80,6 +80,7 @@ Development
 - Add more accessor methods to ``TerkinConfiguration``
 - Add ``get_last_stacktrace`` utility function
 - Fix memory exhaustion when starting the MicroWebSrv twice
+- HTTP API: Add endpoints for getting and setting individual configuration settings
 
 
 2019-06-22 0.5.1